OnWatch/2.5.0/python/requirements.txt
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
urllib3==1.26.9
websocket-client==1.3.1
